Migrating project to use go module support with go 1.11.
A go.mod was created from the existing go dep files. Also, the
docker build is now a two step process and the container is a
minimal container. A static and cross-compiled binary must
first be created on the local machine. This binary is copied into
the container when it is built.